Slate - IT'S FIXED!  No more clip rounding error in Slate causing the last column or row of pixels to sometimes get clipped incorrectly.  In slate we do pixel snapping on each vertex individually, which means that a box that's at x.3 and extends to w.7, will visually rendered at x to (w+1).  The clipping rect was determined in pre-vertex rounding space, the clip rect is represented as origin, extentsX and extentsY.  The extents are all computed in pre-pixel snapping space, so a box that goes from x.3 to w.7, becomes an extents of z.4, and is then rounded down.  Meaning that on certain pixel offsets, some boxes ended up with an extents that is 1 pixel smaller from the visual version where each vertex is rounded individually.  The solution is to perform pixel snapping on the axis aligned clipping rect, which is then used to create the rotated clip rect which is formed with the extents, then to transform it, and again pixel snap.
Also removing the option to use halfs for clipping data when sending it to the pixel shader as it doesn't make a ton of sense to keep this codepath available after discussing it with the others working on Slate.

#codereview matt.kuhlenschmidt nick.atamas wes.hunt

[CL 2583548 by Nick Darnell in Main branch]

+++ Engine/Shaders/StandaloneRenderer/D3D/SlateDefaultVertexShader.hlsl
@@ -17,10 +17,10 @@ struct VertexOut
 };
 
 VertexOut Main(
-	in int2 InPosition : POSITION,
+	in float2 InPosition : POSITION,
 	in float4 InTextureCoordinates : TEXCOORD0,
-	in half2 InClipOrigin : TEXCOORD1,
-	in half4 InClipExtents : TEXCOORD2,
+	in float2 InClipOrigin : TEXCOORD1,
+	in float4 InClipExtents : TEXCOORD2,
 	in float4 InColor : COLOR0
 	) 
 {
